views: decentralise queries into the views
* Each view must register a subscription requesting the
  data the view requires.
* Currently some of these queries are colocated and have not
  been kept in sync with the requirements of the views.
* This commit moves the queries into the views they are used by.
