+++ src/views/Tree.vue
@@ -40,7 +40,138 @@ import graphqlMixin from '@/mixins/graphql'
 import subscriptionComponentMixin from '@/mixins/subscriptionComponent'
 import SubscriptionQuery from '@/model/SubscriptionQuery.model'
 import TreeComponent from '@/components/cylc/tree/Tree.vue'
-import { WORKFLOW_TREE_DELTAS_SUBSCRIPTION } from '@/graphql/queries'
+import gql from 'graphql-tag'
+
+const QUERY = gql`
+subscription Workflow ($workflowId: ID) {
+  deltas (workflows: [$workflowId]) {
+   ...Deltas
+  }
+}
+
+fragment Deltas on Deltas {
+  id
+  added {
+    ...AddedDelta
+  }
+  updated (stripNull: true) {
+    ...UpdatedDelta
+  }
+  pruned {
+    ...PrunedDelta
+  }
+}
+
+fragment AddedDelta on Added {
+  workflow {
+    ...WorkflowData
+  }
+  cyclePoints: familyProxies (ids: ["*/root"]) {
+    ...CyclePointData
+  }
+  familyProxies {
+    ...FamilyProxyData
+  }
+  taskProxies {
+    ...TaskProxyData
+  }
+  jobs {
+    ...JobData
+  }
+}
+
+fragment UpdatedDelta on Updated {
+  taskProxies {
+    ...TaskProxyData
+  }
+  jobs {
+    ...JobData
+  }
+  familyProxies {
+    ...FamilyProxyData
+  }
+}
+
+fragment PrunedDelta on Pruned {
+  familyProxies
+  taskProxies
+  jobs
+}
+
+fragment WorkflowData on Workflow {
+  id
+  status
+  statusMsg
+  owner
+  host
+  port
+  stateTotals
+  latestStateTasks(states: [
+    "failed",
+    "preparing",
+    "submit-failed",
+    "submitted",
+    "running"
+  ])
+}
+
+fragment CyclePointData on FamilyProxy {
+  __typename
+  id
+  state
+  ancestors {
+    name
+  }
+  childTasks {
+    id
+  }
+}
+
+fragment FamilyProxyData on FamilyProxy {
+  __typename
+  id
+  state
+  ancestors {
+    name
+  }
+  childTasks {
+    id
+  }
+}
+
+fragment TaskProxyData on TaskProxy {
+  id
+  state
+  isHeld
+  isQueued
+  isRunahead
+  task {
+    meanElapsedTime
+  }
+  firstParent {
+    id
+  }
+}
+
+fragment JobData on Job {
+  id
+  jobRunnerName
+  jobId
+  platform
+  startedTime
+  submittedTime
+  finishedTime
+  state
+  submitNum
+  messages
+  taskProxy {
+    outputs (satisfied: true) {
+      label
+      message
+    }
+  }
+}
+`
 
 export default {
   name: 'Tree',
@@ -71,7 +202,7 @@ export default {
     },
     query () {
       return new SubscriptionQuery(
-        WORKFLOW_TREE_DELTAS_SUBSCRIPTION,
+        QUERY,
         this.variables,
         'workflow',
         [],
